+++ problemas/iniciante/1011/1011.js
@@ -1,7 +1,7 @@
 let input = require("fs").readFileSync("/dev/stdin", "utf8");
 let lines = input.split("\n");
 
-let R = parseInt(lines.shift());
+let R = parseFloat(lines.shift());
 const PI = 3.14159;
 
 let volume = (4 / 3) * PI * R * R * R;
